refactor(pixel): keep terminal reply plumbing internal
Share graphics pacing and capability probing without exposing their raw TTY
source or scanner across the binary boundary. The sidebar surface ratchet grows
only for the opaque pacer and typed Zellij capability that the CLI consumes.
